Adding docker support

+++ run_all_stages.sh
@@ -6,6 +6,7 @@ set -e  # Exit on error
 TEST_MODE=false
 STAGE=""
 DOCKER_MODE=false
+NO_DOCKER_FLAG=false
 PASSTHRU_ARGS=()
 
 while [[ $# -gt 0 ]]; do
@@ -24,6 +25,10 @@ while [[ $# -gt 0 ]]; do
             DOCKER_MODE=true
             shift
             ;;
+        --no-docker)
+            NO_DOCKER_FLAG=true
+            shift
+            ;;
         *)
             echo "Unknown option: $1"
             echo "Usage: $0 [--test] [--stage N] [--docker]"
@@ -35,11 +40,11 @@ while [[ $# -gt 0 ]]; do
     esac
 done
 
-if [ "$DOCKER_MODE" = true ]; then
+if [ "$DOCKER_MODE" = true ] && [ "$NO_DOCKER_FLAG" = false ]; then
     docker run --rm -v "$(pwd)":/workspace -w /workspace -e IGP_CORES="${IGP_CORES:-}" \
         epi-syrian-refugees ./link_data.sh
     docker run --rm -v "$(pwd)":/workspace -w /workspace -e IGP_CORES="${IGP_CORES:-}" \
-        epi-syrian-refugees ./run_all_stages.sh "${PASSTHRU_ARGS[@]}"
+        epi-syrian-refugees bash -lc "cd /workspace && ./run_all_stages.sh ${PASSTHRU_ARGS[*]} --no-docker"
     exit $?
 fi
